split into client etc
